separate managed and unmanaged code
