+++ pkg/git/libgit2/transport_test.go
@@ -20,7 +20,6 @@ import (
 	"bytes"
 	"context"
 	"crypto/x509"
-	"encoding/base64"
 	"encoding/pem"
 	"errors"
 	"fmt"
@@ -205,159 +204,6 @@ func Test_x509Callback(t *testing.T) {
 	}
 }
 
-func Test_knownHostsCallback(t *testing.T) {
-	tests := []struct {
-		name         string
-		host         string
-		expectedHost string
-		knownHosts   []byte
-		hostkey      git2go.HostkeyCertificate
-		want         error
-	}{
-		{
-			name:         "Match",
-			host:         "github.com",
-			knownHosts:   []byte(knownHostsFixture),
-			hostkey:      git2go.HostkeyCertificate{Kind: git2go.HostkeySHA1 | git2go.HostkeyMD5, HashSHA1: sha1Fingerprint("v2toJdKXfFEaR1u++4iq1UqSrHM")},
-			expectedHost: "github.com",
-			want:         nil,
-		},
-		{
-			name:         "Match with port",
-			host:         "github.com",
-			knownHosts:   []byte(knownHostsFixture),
-			hostkey:      git2go.HostkeyCertificate{Kind: git2go.HostkeySHA1 | git2go.HostkeyMD5, HashSHA1: sha1Fingerprint("v2toJdKXfFEaR1u++4iq1UqSrHM")},
-			expectedHost: "github.com:22",
-			want:         nil,
-		},
-		{
-			name:         "Hostname mismatch",
-			host:         "github.com",
-			knownHosts:   []byte(knownHostsFixture),
-			hostkey:      git2go.HostkeyCertificate{Kind: git2go.HostkeySHA1 | git2go.HostkeyMD5, HashSHA1: sha1Fingerprint("v2toJdKXfFEaR1u++4iq1UqSrHM")},
-			expectedHost: "example.com",
-			want:         fmt.Errorf("host mismatch: %q %q", "example.com", "github.com"),
-		},
-		{
-			name:         "Hostkey mismatch",
-			host:         "github.com",
-			knownHosts:   []byte(knownHostsFixture),
-			hostkey:      git2go.HostkeyCertificate{Kind: git2go.HostkeyMD5, HashMD5: md5Fingerprint("\xb6\x03\x0e\x39\x97\x9e\xd0\xe7\x24\xce\xa3\x77\x3e\x01\x42\x09")},
-			expectedHost: "github.com",
-			want:         fmt.Errorf("hostkey could not be verified"),
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			g := NewWithT(t)
-
-			cert := &git2go.Certificate{Hostkey: tt.hostkey}
-			callback := knownHostsCallback(tt.expectedHost, tt.knownHosts)
-			result := g.Expect(callback(cert, false, tt.host))
-			if tt.want == nil {
-				result.To(BeNil())
-			} else {
-				result.To(Equal(tt.want))
-			}
-		})
-	}
-}
-
-func Test_parseKnownHosts_matches(t *testing.T) {
-	tests := []struct {
-		name        string
-		hostkey     git2go.HostkeyCertificate
-		wantMatches bool
-	}{
-		{"good sha256 hostkey", git2go.HostkeyCertificate{Kind: git2go.HostkeySHA256 | git2go.HostkeySHA1 | git2go.HostkeyMD5, HashSHA256: sha256Fingerprint("nThbg6kXUpJWGl7E1IGOCspRomTxdCARLviKw6E5SY8")}, true},
-		{"bad sha256 hostkey", git2go.HostkeyCertificate{Kind: git2go.HostkeySHA256 | git2go.HostkeySHA1 | git2go.HostkeyMD5, HashSHA256: sha256Fingerprint("ROQFvPThGrW4RuWLoL9tq9I9zJ42fK4XywyRtbOz/EQ")}, false},
-		{"good sha1 hostkey", git2go.HostkeyCertificate{Kind: git2go.HostkeySHA1 | git2go.HostkeyMD5, HashSHA1: sha1Fingerprint("v2toJdKXfFEaR1u++4iq1UqSrHM")}, true},
-		{"bad sha1 hostkey", git2go.HostkeyCertificate{Kind: git2go.HostkeySHA1 | git2go.HostkeyMD5, HashSHA1: sha1Fingerprint("tfpLlQhDDFP3yGdewTvHNxWmAdk")}, false},
-		{"good md5 hostkey", git2go.HostkeyCertificate{Kind: git2go.HostkeyMD5, HashMD5: md5Fingerprint("\x16\x27\xac\xa5\x76\x28\x2d\x36\x63\x1b\x56\x4d\xeb\xdf\xa6\x48")}, true},
-		{"bad md5 hostkey", git2go.HostkeyCertificate{Kind: git2go.HostkeyMD5, HashMD5: md5Fingerprint("\xb6\x03\x0e\x39\x97\x9e\xd0\xe7\x24\xce\xa3\x77\x3e\x01\x42\x09")}, false},
-		{"invalid hostkey", git2go.HostkeyCertificate{}, false},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			g := NewWithT(t)
-
-			knownKeys, err := parseKnownHosts(knownHostsFixture)
-			if err != nil {
-				t.Error(err)
-				return
-			}
-			matches := knownKeys[0].matches("github.com", tt.hostkey)
-			g.Expect(matches).To(Equal(tt.wantMatches))
-		})
-	}
-}
-
-func Test_parseKnownHosts(t *testing.T) {
-	tests := []struct {
-		name    string
-		fixture string
-		wantErr bool
-	}{
-		{
-			name:    "empty file",
-			fixture: "",
-			wantErr: false,
-		},
-		{
-			name:    "single host",
-			fixture: `github.com ssh-rsa AAAAB3NzaC1yc2EAAAABIwAAAQEAq2A7hRGmdnm9tUDbO9IDSwBK6TbQa+PXYPCPy6rbTrTtw7PHkccKrpp0yVhp5HdEIcKr6pLlVDBfOLX9QUsyCOV0wzfjIJNlGEYsdlLJizHhbn2mUjvSAHQqZETYP81eFzLQNnPHt4EVVUh7VfDESU84KezmD5QlWpXLmvU31/yMf+Se8xhHTvKSCZIFImWwoG6mbUoWf9nzpIoaSjB+weqqUUmpaaasXVal72J+UX2B+2RPW3RcT0eOzQgqlJL3RKrTJvdsjE3JEAvGq3lGHSZXy28G3skua2SmVi/w4yCE6gbODqnTWlg7+wC604ydGXA8VJiS5ap43JXiUFFAaQ==`,
-			wantErr: false,
-		},
-		{
-			name: "single host with comment",
-			fixture: `# github.com
-github.com ssh-rsa AAAAB3NzaC1yc2EAAAABIwAAAQEAq2A7hRGmdnm9tUDbO9IDSwBK6TbQa+PXYPCPy6rbTrTtw7PHkccKrpp0yVhp5HdEIcKr6pLlVDBfOLX9QUsyCOV0wzfjIJNlGEYsdlLJizHhbn2mUjvSAHQqZETYP81eFzLQNnPHt4EVVUh7VfDESU84KezmD5QlWpXLmvU31/yMf+Se8xhHTvKSCZIFImWwoG6mbUoWf9nzpIoaSjB+weqqUUmpaaasXVal72J+UX2B+2RPW3RcT0eOzQgqlJL3RKrTJvdsjE3JEAvGq3lGHSZXy28G3skua2SmVi/w4yCE6gbODqnTWlg7+wC604ydGXA8VJiS5ap43JXiUFFAaQ==`,
-			wantErr: false,
-		},
-		{
-			name: "multiple hosts with comments",
-			fixture: `# github.com
-github.com ssh-rsa AAAAB3NzaC1yc2EAAAABIwAAAQEAq2A7hRGmdnm9tUDbO9IDSwBK6TbQa+PXYPCPy6rbTrTtw7PHkccKrpp0yVhp5HdEIcKr6pLlVDBfOLX9QUsyCOV0wzfjIJNlGEYsdlLJizHhbn2mUjvSAHQqZETYP81eFzLQNnPHt4EVVUh7VfDESU84KezmD5QlWpXLmvU31/yMf+Se8xhHTvKSCZIFImWwoG6mbUoWf9nzpIoaSjB+weqqUUmpaaasXVal72J+UX2B+2RPW3RcT0eOzQgqlJL3RKrTJvdsjE3JEAvGq3lGHSZXy28G3skua2SmVi/w4yCE6gbODqnTWlg7+wC604ydGXA8VJiS5ap43JXiUFFAaQ==
-# gitlab.com
-gitlab.com ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIAfuCHKVTjquxvt6CM6tdG4SLp1Btn/nOeHHE5UOzRdf`,
-		},
-		{
-			name: "no host key, only comments",
-			fixture: `# example.com
-#github.com
-# gitlab.com`,
-			wantErr: false,
-		},
-		{
-			name:    "invalid host entry",
-			fixture: `github.com ssh-rsa`,
-			wantErr: true,
-		},
-		{
-			name:    "invalid content",
-			fixture: `some random text`,
-			wantErr: true,
-		},
-		{
-			name: "invalid line with valid host key",
-			fixture: `some random text
-gitlab.com ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIAfuCHKVTjquxvt6CM6tdG4SLp1Btn/nOeHHE5UOzRdf`,
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			g := NewWithT(t)
-			_, err := parseKnownHosts(tt.fixture)
-			if tt.wantErr {
-				g.Expect(err).To(HaveOccurred())
-			} else {
-				g.Expect(err).ToNot(HaveOccurred())
-			}
-		})
-	}
-}
-
 func Test_transferProgressCallback(t *testing.T) {
 	tests := []struct {
 		name       string
@@ -520,94 +366,6 @@ func Test_pushTransferProgressCallback(t *testing.T) {
 			}
 		})
 	}
-}
-
-func TestMatchHashedHost(t *testing.T) {
-	tests := []struct {
-		name      string
-		knownHost string
-		host      string
-		match     bool
-		wantErr   string
-	}{
-		{
-			name:      "match valid known host",
-			knownHost: "|1|vApZG0Ybr4rHfTb69+cjjFIGIv0=|M5sSXen14encOvQAy0gseRahnJw=",
-			host:      "[127.0.0.1]:44167",
-			match:     true,
-		},
-		{
-			name:    "empty known host errors",
-			wantErr: "hashed known host must begin with '|'",
-		},
-		{
-			name:      "unhashed known host errors",
-			knownHost: "[127.0.0.1]:44167",
-			wantErr:   "hashed known host must begin with '|'",
-		},
-		{
-			name:      "invalid known host format errors",
-			knownHost: "|1M5sSXen14encOvQAy0gseRahnJw=",
-			wantErr:   "invalid format for hashed known host",
-		},
-		{
-			name:      "invalid hash type errors",
-			knownHost: "|2|vApZG0Ybr4rHfTb69+cjjFIGIv0=|M5sSXen14encOvQAy0gseRahnJw=",
-			wantErr:   "unsupported hash type",
-		},
-		{
-			name:      "invalid base64 component[2] errors",
-			knownHost: "|1|azz|M5sSXen14encOvQAy0gseRahnJw=",
-			wantErr:   "cannot decode hashed known host",
-		},
-		{
-			name:      "invalid base64 component[3] errors",
-			knownHost: "|1|M5sSXen14encOvQAy0gseRahnJw=|azz",
-			wantErr:   "cannot decode hashed known host",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			g := NewWithT(t)
-
-			matched, err := MatchHashedHost(tt.knownHost, tt.host)
-
-			if tt.wantErr == "" {
-				g.Expect(err).NotTo(HaveOccurred())
-				g.Expect(matched).To(Equal(tt.match))
-			} else {
-				g.Expect(err).To(HaveOccurred())
-				g.Expect(err.Error()).To(ContainSubstring(tt.wantErr))
-			}
-		})
-	}
-}
-
-func md5Fingerprint(in string) [16]byte {
-	var out [16]byte
-	copy(out[:], in)
-	return out
-}
-
-func sha1Fingerprint(in string) [20]byte {
-	d, err := base64.RawStdEncoding.DecodeString(in)
-	if err != nil {
-		panic(err)
-	}
-	var out [20]byte
-	copy(out[:], d)
-	return out
-}
-
-func sha256Fingerprint(in string) [32]byte {
-	d, err := base64.RawStdEncoding.DecodeString(in)
-	if err != nil {
-		panic(err)
-	}
-	var out [32]byte
-	copy(out[:], d)
-	return out
 }
 
 func certificateFromPEM(pemBytes string) (*x509.Certificate, error) {
